+++ src/types/number.rs
@@ -9,6 +9,9 @@ use float_cmp::ApproxEqUlps;
 
 use {WriteOptions, WriteBuffer};
 
+/// A default numeric precision.
+pub const DEFAULT_PRECISION: u8 = 12;
+
 /// The trait for comparing f64 numbers.
 pub trait FuzzyEq {
     /// Returns `true` if numbers are equal.
@@ -47,13 +50,33 @@ impl FuzzyOrd for f64 {
     }
 }
 
-pub fn write_num(num: f64, rm_leading_zero: bool, buf: &mut Vec<u8>) {
-    // We always round numbers up to 12 digits
+static POW_VEC: &'static [f64] = &[
+                    0.0,
+                   10.0,
+                  100.0,
+                1_000.0,
+               10_000.0,
+              100_000.0,
+            1_000_000.0,
+           10_000_000.0,
+          100_000_000.0,
+        1_000_000_000.0,
+       10_000_000_000.0,
+      100_000_000_000.0,
+    1_000_000_000_000.0,
+];
+
+pub fn write_num(num: &f64, precision: u8, rm_leading_zero: bool, buf: &mut Vec<u8>) {
+    debug_assert!(precision > 0 && precision <= 12);
+
+    let prec = precision as usize;
+
+    // By default it will round numbers up to 12 digits
     // to prevent writing ugly numbers like 29.999999999999996.
     // It's not 100% correct, but differences are insignificant.
 
     // TODO: find a faster/more correct method
-    let v = (num * 1_000_000_000_000.0f64).round() / 1_000_000_000_000.0f64;
+    let v = (num * POW_VEC[prec]).round() / POW_VEC[prec];
 
     let start_pos = buf.len();
 
@@ -84,7 +107,7 @@ pub fn write_num(num: f64, rm_leading_zero: bool, buf: &mut Vec<u8>) {
 
 impl WriteBuffer for f64 {
     fn write_buf_opt(&self, opt: &WriteOptions, buf: &mut Vec<u8>) {
-        write_num(*self, opt.remove_leading_zero, buf);
+        write_num(self, DEFAULT_PRECISION, opt.remove_leading_zero, buf);
     }
 }
 
@@ -97,21 +120,21 @@ mod tests {
             #[test]
             fn $name() {
                 let mut v = Vec::new();
-                write_num($num, $rm_zero, &mut v);
+                write_num(&$num, DEFAULT_PRECISION, $rm_zero, &mut v);
                 assert_eq!(String::from_utf8(v).unwrap(), $result);
             }
         )
     }
 
-    test_number!(gen_number_1, 1.0,                  false, "1");
-    test_number!(gen_number_2, 0.0,                  false, "0");
-    test_number!(gen_number_3, -0.0,                 false, "0");
-    test_number!(gen_number_4, -1.0,                 false, "-1");
-    test_number!(gen_number_5, 12345678.12345678,    false, "12345678.12345678");
-    test_number!(gen_number_6, -0.1,                 true,  "-.1");
-    test_number!(gen_number_7, 0.1,                  true,  ".1");
-    test_number!(gen_number_8, 1.0,                  true,  "1");
-    test_number!(gen_number_9, -1.0,                 true,  "-1");
+    test_number!(gen_number_1,  1.0,                 false, "1");
+    test_number!(gen_number_2,  0.0,                 false, "0");
+    test_number!(gen_number_3,  -0.0,                false, "0");
+    test_number!(gen_number_4,  -1.0,                false, "-1");
+    test_number!(gen_number_5,  12345678.12345678,   false, "12345678.12345678");
+    test_number!(gen_number_6,  -0.1,                true,  "-.1");
+    test_number!(gen_number_7,  0.1,                 true,  ".1");
+    test_number!(gen_number_8,  1.0,                 true,  "1");
+    test_number!(gen_number_9,  -1.0,                true,  "-1");
     test_number!(gen_number_10, 1.5,                 false, "1.5");
     test_number!(gen_number_11, 0.14186,             false, "0.14186");
     test_number!(gen_number_12, 29.999999999999996,  false, "30");